LSRC-717 LST should handle far-North scenes - fixed by expanding the data type from 8-bit to 16-bit for the grid index (which is a row x column value).

+++ not-validated-prototype_lst/src/lst_types.h
@@ -5,7 +5,7 @@
 #include <limits.h>
 
 typedef struct {
-    int8_t index;
+    int16_t index;
     int8_t run_modtran;
     uint8_t row;
     uint8_t col;
